quantum-crypto-bot/requirements.txt
pyTelegramBotAPI==4.14.1
requests==2.31.0
python-dotenv==1.0.0
schedule==1.2.0
pytz==2023.3
matplotlib==3.7.0
pandas==2.0.0